+++ src/softaware.Cqs.Tests/CQ.Handlers/CommandHandlers/ValidationCommandHandler.cs
@@ -1,12 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using softaware.Cqs.Tests.CQ.Contract.Commands;
 
 namespace softaware.Cqs.Tests.CQ.Handlers.CommandHandlers
 {
-    internal class ValidationCommandHandler : ICommandHandler<ValidationCommand>
+    public class ValidationCommandHandler : ICommandHandler<ValidationCommand>
     {
         public Task HandleAsync(ValidationCommand command)
         {