+++ src/softaware.Cqs.Tests/CommandAndQueryProcessorTest.cs
@@ -1,21 +1,16 @@
+using System;
 using System.Reflection;
 using System.Threading.Tasks;
+using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
 using SimpleInjector;
-using softaware.Cqs.SimpleInjector;
 using softaware.Cqs.Tests.CQ.Contract.Commands;
 using softaware.Cqs.Tests.CQ.Contract.Queries;
 
 namespace softaware.Cqs.Tests
 {
-    public class CommandAndQueryProcessorTest : TestBase
+    public abstract class CommandAndQueryProcessorTest : TestBase
     {
-        public override void SetUp()
-        {
-            base.SetUp();
-            this.RegisterPublicDecoratorsAndVerifyContainer();
-        }
-
         [Test]
         public async Task ExecuteCommand()
         {
@@ -34,6 +29,52 @@ namespace softaware.Cqs.Tests
             var result = await this.queryProcessor.ExecuteAsync(query);
 
             Assert.That(result, Is.EqualTo(16));
+        }
+
+        private class SimpleInjectorTest
+            : CommandAndQueryProcessorTest
+        {
+            private Container container;
+
+            [SetUp]
+            public override void SetUp()
+            {
+                this.container = new Container();
+
+                this.container
+                    .AddSoftawareCqs(b => b.IncludeTypesFrom(Assembly.GetExecutingAssembly()))
+                    .AddDecorators(b => { });
+
+                this.container.Verify();
+
+                base.SetUp();
+            }
+
+            protected override ICommandProcessor GetCommandProcessor() => this.container.GetRequiredService<ICommandProcessor>();
+            protected override IQueryProcessor GetQueryProcessor() => this.container.GetRequiredService<IQueryProcessor>();
+        }
+
+        private class ServiceCollectionTest
+            : CommandAndQueryProcessorTest
+        {
+            private IServiceProvider serviceProvider;
+
+            [SetUp]
+            public override void SetUp()
+            {
+                var services = new ServiceCollection();
+
+                services
+                    .AddSoftawareCqs(b => b.IncludeTypesFrom(Assembly.GetExecutingAssembly()))
+                    .AddDecorators(b => { });
+
+                this.serviceProvider = services.BuildServiceProvider();
+
+                base.SetUp();
+            }
+
+            protected override ICommandProcessor GetCommandProcessor() => this.serviceProvider.GetRequiredService<ICommandProcessor>();
+            protected override IQueryProcessor GetQueryProcessor() => this.serviceProvider.GetRequiredService<IQueryProcessor>();
         }
     }
 }