+++ src/softaware.Cqs.Tests/TransactionDecoratorTest.cs
@@ -1,29 +1,19 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
+using System.Reflection;
 using System.Threading.Tasks;
 using System.Transactions;
+using Microsoft.Extensions.DependencyInjection;
 using NUnit.Framework;
+using SimpleInjector;
 using softaware.Cqs.Decorators.Transaction;
-using softaware.Cqs.SimpleInjector;
 using softaware.Cqs.Tests.CQ.Contract.Commands;
 using softaware.Cqs.Tests.CQ.Contract.Queries;
 
 namespace softaware.Cqs.Tests
 {
     [TestFixture]
-    public class TransactionDecoratorTest : TestBase
+    public abstract class TransactionDecoratorTest : TestBase
     {
-        public override void SetUp()
-        {
-            base.SetUp();
-
-            container.RegisterDecorator(typeof(ICommandHandler<>), typeof(TransactionAwareCommandHandlerDecorator<>));
-            container.RegisterDecorator(typeof(IQueryHandler<,>), typeof(TransactionAwareQueryHandlerDecorator<,>));
-
-            this.RegisterPublicDecoratorsAndVerifyContainer();
-        }
-
         [Test]
         public async Task TestCommandTransaction_CommandDoesNotThrow_CommitsTransaction()
         {
@@ -104,6 +94,56 @@ namespace softaware.Cqs.Tests
             Assert.That(exception.Message, Is.EqualTo("We throw here for testing the rollback of transactions."));
 
             Assert.That(transactionCommitted, Is.False);
+        }
+
+        private class SimpleInjectorTest
+            : TransactionDecoratorTest
+        {
+            private Container container;
+
+            [SetUp]
+            public override void SetUp()
+            {
+                this.container = new Container();
+
+                this.container
+                    .AddSoftawareCqs(b => b.IncludeTypesFrom(Assembly.GetExecutingAssembly()))
+                    .AddDecorators(b => b
+                        .AddQueryHandlerDecorator(typeof(TransactionAwareQueryHandlerDecorator<,>))
+                        .AddCommandHandlerDecorator(typeof(TransactionAwareCommandHandlerDecorator<>)));
+
+                this.container.Verify();
+
+                base.SetUp();
+            }
+
+            protected override ICommandProcessor GetCommandProcessor() => this.container.GetRequiredService<ICommandProcessor>();
+            protected override IQueryProcessor GetQueryProcessor() => this.container.GetRequiredService<IQueryProcessor>();
+        }
+
+        private class ServiceCollectionTest
+            : TransactionDecoratorTest
+        {
+            private IServiceProvider serviceProvider;
+
+            [SetUp]
+            public override void SetUp()
+            {
+                var services = new ServiceCollection();
+
+                services
+                    .AddSoftawareCqs(b => b.IncludeTypesFrom(Assembly.GetExecutingAssembly()))
+                    .AddDecorators(b => b
+                        .AddTransactionCommandHandlerDecorator()
+                        .AddTransactionQueryHandlerDecorator());
+
+                this.serviceProvider = services.BuildServiceProvider();
+
+                base.SetUp();
+            }
+
+            protected override ICommandProcessor GetCommandProcessor() => this.serviceProvider.GetRequiredService<ICommandProcessor>();
+            protected override IQueryProcessor GetQueryProcessor() => this.serviceProvider.GetRequiredService<IQueryProcessor>();
         }
     }
 }